Move mocks to separate files. Add duplicate semesters error and test.
